Extract frontmatter logic to own class
